+++ CHANGELOG.md
@@ -1,3 +1,22 @@
-## 0.0.1
+## [0.0.1] - 2023-02-17
 
-* TODO: Describe initial release.
+* Start your app from scratch.
+
+* How to use it?
+* 
+* main.dart
+*  
+* wrap MyApp() in runApp
+*  
+* [FatiYazdiState] runApp(const FatihState(child: MyApp()));
+*  
+* Import it
+* Now in your Dart code, you can use:
+*  
+* [import '../fati_yazdi_state.dart';]
+*  
+* Usage
+*  
+* call from anywhere
+* [FatiYazdiState.fatihStateCreateAgain] FatihState.fatihStateCreateAgain(context);
+